Crate disk_initializer

Disk initializer application

This app is in charge of generating a valid disk image from a file name, a size and a template directory (containing files that should be on the disk image)

The disk image contains a standard GPT partition layout and a FAT filesystem. From the disk size, it chooses a FAT filesystem that may be appropriate to fit in it.

Usage: <disk_name> <disk_size> <template_path>

Modules

gpt
GPT definition module.

Constants

BLOCK_SIZE
Size of a block.

BLOCK_SIZE_U64
Size of an AHCI block in u64.

Functions

main

write_file_to_filesystem
Write a std file to FAT filesystem.

write_tempate_to_filesystem
Write the template directory content to a FAT fileystem